requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
elasticsearch==8.13.0
pdfplumber==0.10.4
python-docx==1.1.0
striprtf==0.0.28
Werkzeug==3.0.3
gunicorn==22.0.0